Body Dimensions

Body girth measurements and skeletal diameter measurements, as well as age, weight, height and gender, are given for 507 physically active individuals - 247 men and 260 women.

bdims

Format

A tbl_df with 507 rows and 25 variables.

bia.di

A numerical vector, respondent's biacromial diameter in centimeters.

bii.di

A numerical vector, respondent's biiliac diameter (pelvic breadth) in centimeters.

bit.di

A numerical vector, respondent's bitrochanteric diameter in centimeters.

che.de

A numerical vector, respondent's chest depth in centimeters, measured between spine and sternum at nipple level, mid-expiration.

che.di

A numerical vector, respondent's chest diameter in centimeters, measured at nipple level, mid-expiration.

elb.di

A numerical vector, respondent's elbow diameter in centimeters, measured as sum of two elbows.

wri.di

A numerical vector, respondent's wrist diameter in centimeters, measured as sum of two wrists.

kne.di

A numerical vector, respondent's knee diameter in centimeters, measured as sum of two knees.

ank.di

A numerical vector, respondent's ankle diameter in centimeters, measured as sum of two ankles.

sho.gi

A numerical vector, respondent's shoulder girth in centimeters, measured over deltoid muscles.

che.gi

A numerical vector, respondent's chest girth in centimeters, measured at nipple line in males and just above breast tissue in females, mid-expiration.

wai.gi

A numerical vector, respondent's waist girth in centimeters, measured at the narrowest part of torso below the rib cage as average of contracted and relaxed position.

nav.gi

A numerical vector, respondent's navel (abdominal) girth in centimeters, measured at umbilicus and iliac crest using iliac crest as a landmark.

hip.gi

A numerical vector, respondent's hip girth in centimeters, measured at at level of bitrochanteric diameter.

thi.gi

A numerical vector, respondent's thigh girth in centimeters, measured below gluteal fold as the average of right and left girths.

bic.gi

A numerical vector, respondent's bicep girth in centimeters, measured when flexed as the average of right and left girths.

for.gi

A numerical vector, respondent's forearm girth in centimeters, measured when extended, palm up as the average of right and left girths.

kne.gi

A numerical vector, respondent's knee diameter in centimeters, measured as sum of two knees.

cal.gi

A numerical vector, respondent's calf maximum girth in centimeters, measured as average of right and left girths.

ank.gi

A numerical vector, respondent's ankle minimum girth in centimeters, measured as average of right and left girths.

wri.gi

A numerical vector, respondent's wrist minimum girth in centimeters, measured as average of right and left girths.

age

A numerical vector, respondent's age in years.

wgt

A numerical vector, respondent's weight in kilograms.

hgt

A numerical vector, respondent's height in centimeters.

sex

A categorical vector, "m" if the respondent is male, "f" if female.

Source

Heinz G, Peterson LJ, Johnson RW, Kerk CJ. 2003. Exploring Relationships in Body Dimensions. Journal of Statistics Education 11(2).