lab_report

Acts as a simplified template to common parameters passed to rmarkdown::html_document

lab_report(highlight = "pygments", theme = "spacelab", toc = TRUE,
  toc_float = TRUE, code_download = TRUE, code_folding = "show")

Arguments

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "textmate". Pass NULL to prevent syntax highlighting.

theme
Visual theme ("default", "cerulean", "journal", "flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper", "sandstone", "simplex", or "yeti"). Pass NULL for no theme (in this case you can use the css parameter to add your own styles).

toc
TRUE to include a table of contents in the output

toc_float
TRUE to float the table of contents to the left of the main document content. Rather than TRUE you may also pass a list of options that control the behavior of the floating table of contents. See the Floating Table of Contents section below for details.

code_download
Embed the Rmd source code within the document and provide a link that can be used by readers to download the code.

code_folding
Enable document readers to toggle the display of R code chunks. Specify "none" to display all code chunks (assuming they were knit with echo = TRUE). Specify "hide" to hide all R code chunks by default (users can show hidden code chunks either individually or document-wide). Specify "show" to show all R code chunks by default.